+++ app/app.py
@@ -2,6 +2,8 @@ from flask import Flask
 from flask_migrate import Migrate
 from flask_cors import CORS
 from main import main_bp
+from billing_info import billing_info_bp
+from tags import tags_bp
 import os
 
 
@@ -17,6 +19,8 @@ def create_app():
     migrate = Migrate(app, db)
     app.register_blueprint(main_bp)
     app.register_blueprint(auth_bp)
+    app.register_blueprint(billing_info_bp)
+    app.register_blueprint(tags_bp)
     jwt.init_app(app)
     bcrypt.init_app(app)
     CORS(app, resources={r"*": {"origins": "*"}})
